+++ inst/doc/zeropower.R
@@ -1,4 +1,4 @@
-## ----setup, include=FALSE-----------------------------------------------------
+## ----setup, include=FALSE---------------------------------
 set.seed(0)
 library("hyper2")
 library("magrittr")
@@ -13,7 +13,7 @@ registerS3method(
 ## ----out.width='20%', out.extra='style="float:right; padding:10px"',echo=FALSE----
 knitr::include_graphics(system.file("help/figures/hyper2.png", package = "hyper2"))
 
-## ----appidiom-----------------------------------------------------------------
+## ----appidiom---------------------------------------------
 chess <- hyper2()
 chess["Topalov"] <- 30
 chess["Anand"  ] <- 36
@@ -22,23 +22,22 @@ chess[c("Topalov","Anand" )] <- 35  # bug!  should  be -35
 chess[c("Anand","Karpov"  )] <- 35  # bug!  should  be -35
 chess[c("Karpov","Topalov")] <- 18  # bug!  should  be -18
 
-## ----showwarningunbalanced----------------------------------------------------
+## ----showwarningunbalanced--------------------------------
 chess
 
-## ----correctid----------------------------------------------------------------
+## ----correctid--------------------------------------------
 chess[c("Topalov","Anand" )] <- -35
 chess[c("Anand","Karpov"  )] <- -35
 chess[c("Karpov","Topalov")] <- -18
 chess
 loglik(equalp(chess),chess)
 
-## ----newinfo------------------------------------------------------------------
-chess["Anand" ] %<>% inc
-chess["Karpov"] %<>% inc
-chess[c("Anand,Karpov")]  %<>% dec(2)
+## ----newinfo----------------------------------------------
+chess["Anand"] %<>% inc
+chess[c("Anand, Karpov")] %<>% dec(2)
 chess
 
-## -----------------------------------------------------------------------------
+## ---------------------------------------------------------
 H <- hyper2()
 H['a'] %<>% inc(5)
 H['b'] %<>% inc(2)