+++ NEWS.md
@@ -1,6 +1,9 @@
 # hyper2 3.1-0
 
 * consistent handling of non-zero-sum powers 
+* no longer imports frab functionality
+* new function dirichlet3()
+* reverse dependency check workflow
 
 # hyper2 3.0-1
 
@@ -20,4 +23,5 @@
 # hyper2 2.0-2
 
 * implement `alabama::constrOptim.nl()` as an alternative to `base::constrOptim()` to avoid wmmin bug
+